+++ types/services/notifier.go
@@ -28,6 +28,6 @@ func FindNotifier(method string) *notifications.Notification {
 type ServiceNotifier interface {
 	OnSuccess(*Service) error                    // OnSuccess is triggered when a service is successful
 	OnFailure(*Service, *failures.Failure) error // OnFailure is triggered when a service is failing
-	OnTest() error                               // OnTest is triggered for testing
+	OnTest() (string, error)                     // OnTest is triggered for testing
 	Select() *notifications.Notification         // OnTest is triggered for testing
 }